requirements.txt
pandas==1.0.5
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
six==1.15.0
xlrd==1.2.0
